requirements.txt
onnxruntime>=1.17.0
pillow>=10.2.0
numpy>=1.26.4
ultralytics==8.1.0 